crawlservice/requirements.txt
py-eureka-client==0.11.13
openpyxl==3.1.1
fastapi
uvicorn
pydantic
facebook-sdk